# Conflicts:
#	redisson/src/main/java/org/redisson/misc/RedissonPromise.java

+++ redisson/src/test/java/org/redisson/executor/ScheduledLongRepeatableTask.java
@@ -23,7 +23,7 @@ public class ScheduledLongRepeatableTask implements Runnable {
     @Override
     public void run() {
         if (redisson.getAtomicLong(counterName).incrementAndGet() == 3) {
-            for (int i = 0; i < Long.MAX_VALUE; i++) {
+            for (long i = 0; i < Long.MAX_VALUE; i++) {
                 if (Thread.currentThread().isInterrupted()) {
                     System.out.println("interrupted " + i);
                     redisson.getBucket(objectName).set(i);
